requirements.txt
streamlit==1.33.0
pdfplumber==0.11.4
spacy==3.7.5
sentence-transformers==2.7.0
torch==2.2.2
transformers==4.41.2
faiss-cpu==1.8.0.post1
ollama==0.2.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
protobuf==3.20.3